+++ cli/cli/commands/service/logs/logs.go
@@ -25,6 +25,8 @@ import (
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"io"
+	"os"
+	"os/signal"
 	"strconv"
 )
 
@@ -129,7 +131,7 @@ func run(
 		}
 		defer func() {
 			for _, userServiceLogsReadCloser := range successfulUserServiceLogs {
-				if err := userServiceLogsReadCloser.Close(); err != nil{
+				if err := userServiceLogsReadCloser.Close(); err != nil {
 					logrus.Warnf("We tried to close the user service logs read-closer-objects after we're done using it, but doing so threw an error:\n%v", err)
 				}
 			}
@@ -175,7 +177,7 @@ func run(
 
 		for _, logLine := range serviceLogs {
 			if _, err := fmt.Fprintln(logrus.StandardLogger().Out, logLine); err != nil {
-				logrus.Errorf("We tried to print the user service log line '%v', but doing so threw an error:\n%v",logLine, err)
+				logrus.Errorf("We tried to print the user service log line '%v', but doing so threw an error:\n%v", logLine, err)
 			}
 		}
 
@@ -188,22 +190,28 @@ func run(
 	}
 	defer cancelStreamUserServiceLogsFunc()
 
+	// This channel will receive a signal when the user presses an interrupt
+	interruptChan := make(chan os.Signal)
+	signal.Notify(interruptChan, os.Interrupt)
+
 	for {
-		userServiceLogsByGuid, isChanOpen := <-userServiceLogsByGuidChan
-		if !isChanOpen {
-			break
-		}
+		select {
+		case userServiceLogsByGuid, isChanOpen := <-userServiceLogsByGuidChan:
+			if !isChanOpen {
+				return nil
+			}
 
-		userServiceLogs, found := userServiceLogsByGuid[serviceGuid]
-		if !found {
-			return stacktrace.NewError("Expected to find logs for user service with GUID '%v' on user service logs map '%+v' but was not found; this should never happen, and is a bug in Kurtosis", serviceGuid, userServiceLogsByGuid)
-		}
+			userServiceLogs, found := userServiceLogsByGuid[serviceGuid]
+			if !found {
+				return stacktrace.NewError("Expected to find logs for user service with GUID '%v' on user service logs map '%+v' but was not found; this should never happen, and is a bug in Kurtosis", serviceGuid, userServiceLogsByGuid)
+			}
 
-		for _, serviceLog := range userServiceLogs {
-			fmt.Println(serviceLog.GetContent())
+			for _, serviceLog := range userServiceLogs {
+				logrus.Println(serviceLog.GetContent())
+			}
+		case <-interruptChan:
+			logrus.Debugf("Received signal interruption in service logs Kurtosis CLI command")
+			return nil
 		}
-
 	}
-
-	return nil
 }